+++ FrensHelpers.h
@@ -30,6 +30,11 @@
 #ifndef PSRAM_CS_PIN
 #define PSRAM_CS_PIN 0 // 0 is no PSRAM
 #endif
+#if !HSTX
+#define FILEXTFORSEARCH ".444"
+#else
+#define FILEXTFORSEARCH ".555"
+#endif
 
 enum class ScreenMode
     {
@@ -66,6 +71,7 @@ extern char __flash_binary_start;  // defined in linker script
 extern char __flash_binary_end; 
 extern int abSwapped;      // defined in hid_app.cpp
 extern int isManta;        // defined in hid_app.cpp
+
 namespace Frens
 {
 #if !HSTX && FRAMEBUFFERISPOSSIBLE
@@ -127,7 +133,9 @@ namespace Frens
     const char* ms_to_d_hhmmss(uint64_t ms, char* buf, size_t bufSize);
     void setClocksAndStartStdio(uint32_t cpuFreqKHz, vreg_voltage voltage);
     //extern volatile ProcessScanLineFunction processScanLineFunction;
-   
+    void loadOverLay(const char *filename, const char *overlay);
+    FRESULT pick_random_file_fullpath(const char *path, char *chosen, size_t bufsize);
+    uint32_t getCrcOfLoadedRom();
 } // namespace Frens